feat: update imports to absolute and redundant file names to index
