+++ config-overrides.js
@@ -1,0 +1,5 @@
+/* eslint-disable no-param-reassign */
+module.exports = function override(config) {
+  config.resolve.modules = ["node_modules", "src"];
+  return config;
+};
