+++ src/components/Die/index.jsx
