+++ src/components/MinigameSelector/index.jsx
