+++ src/components/Star/index.jsx
