+++ src/hooks/useSetRealms.js
@@ -3,8 +3,8 @@ import { useSelector, useDispatch } from "react-redux";
 import {
   setQueryStringValue,
   setRealmsUsingQueryString,
-} from "../utilities/queryString";
-import { actionCreators } from "../reducers/app";
+} from "utilities/queryString";
+import { actionCreators } from "reducers/app";
 
 const useSetRealms = () => {
   const dispatch = useDispatch();
